seed refactoring and files imports changes
